Cleaned JARP Data
The jarp_master.dta file is the cleaned JARP data that contains all 3 generations (Issei, Nisei, Sansei). The file jarp.nisei.dta is the version of the data that contains only the second-generation Japanese Americans. It has the Nisei's parents' income level because I merged that from Issei's data based on family ID. 
The file jarp.do is the do file that I used to clean all the raw files and create jarp_master.dta and jarp.nisei.dta. It also has the code that I used to create tables in excel.